requirements.txt
flask>=3.0
python-dotenv>=1.0
anthropic>=0.42
openai>=1.60
google-generativeai>=0.8
PyPDF2>=3.0
python-docx>=1.1
werkzeug>=3.0
markdown>=3.5
